fix(gateway): preserve personal account pins in model readiness
Project personal defaults and user-link session pins at their existing execution strength. Keep shared catalog snapshots isolated and explicit account selection authoritative.
